Moving top-entity code to module
